11_handoffs/pyproject.toml
[project]
name = "11-handoffs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.26.0",
    "google-generativeai>=0.8.5",
    "openai-agents>=0.2.2",
    "python-dotenv>=1.1.1",
    "rich>=14.0.0",
]
